pyproject.toml
[tool.poetry]
name = "bottica"
version = "0.25.1"
description = "A personal music-playing discord bot project."
authors = ["Technomunk <thegriffones@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.6"
click = "^8.3"
dataclass-csv = "^1.4"
pepver = "^1.1"
sentry-sdk = "^2.25"
toml = "^0.10"

[tool.poetry.scripts]
bottica = "bottica.__main__:cli"

[tool.poetry.group.dev.dependencies]
black = "*"
flake8 = "*"
mypy = "*"
pylint = "*"
pytest = "^9.0"
types-toml = "^0.10"
flake8-pyproject = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.flake8]
extend-ignore = ["E203"]
max-line-length = 120
extend-exclude = "bottica/music"

[tool.mypy]
python_version = "3.10"
explicit_package_bases = true
namespace_packages = true
exclude = ['/bottica/music/']
